backend/pyproject.toml
[tool.poetry]
name = "blog-backend"
version = "0.1.0"
description = ""
authors = ["ajaypawar22112001 <ajay.pawar@reversebits.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.5.0"
pytest = "^8.3.5"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
